Remove Value facets from higher (web app) layer of the SDK.
After a discussion with Roman, I realized the value facets
work differently then I originally thought. They will have
to be designed differently on the web app level, and we have
to decide how exactly.

On the lower (client) level, the Values facets still remain
and map exactly to what the HTTP API does - the provided
values are combined using OR - a sum of all is returned.