Make range facets inclusive on both intervals.
ElasticSearch range facets are always
[lower_inclusive, upper_exclusive), without any possibility
of configuration. This doesn't match how our filters behave:
[lower_inclusive, upper_inclusive].
To get around this, we adjust the upper bound of facet ranges
manually.
For example, if a user defines a facet [0, 10], [10, 20] we
transform these intervals into [0, 10.001), [10, 20.001) for
the facet, they stay unchanged for filters.
As a result, a product with the exact value of 10 will be counted
(by the facet) and returned (by the filter) for exactly the same
intervals, which makes typical multiselect facets behave in a
consistent way.